feat: 暂存请求, 统一保存

+++ internal/model/device.go
@@ -8,7 +8,7 @@ const OauthLoginDeviceMagic = "oauth_login"
 type Device struct {
 	UserID uint   `json:"user_id"`
 	Magic  string `json:"magic"`
-	Count  int    `gorm:"default:0" json:"count"`
+	Count  int64  `gorm:"default:0" json:"count"`
 	BaseModel
 }
